% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_mgh.R
\name{as_mgh}
\alias{as_mgh}
\title{Create an object that is structured like an mgh object}
\usage{
as_mgh(x, v = NULL, ndim1 = NULL, ndim2 = NULL, ndim3 = NULL,
  nframes, type, dof)
}
\arguments{
\item{x}{the vertex-wise values}

\item{v}{Version (default = 1)}

\item{ndim1}{Width / 1st dimension}

\item{ndim2}{Height / 2nd dimension}

\item{ndim3}{Depth / 3rd dimension}

\item{nframes}{Number of scalar components}

\item{type}{Data type, can be UCHAR (0), SHORT (4), INT (1) or FLOAT (3)}

\item{dof}{Degrees of freedom}
}
\description{
Create an object that is structured like an mgh object
}
