% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_mgh.R
\name{bsfbm2mgh}
\alias{bsfbm2mgh}
\title{Save out an MGH file from memory}
\usage{
bsfbm2mgh(fbm, fname, filter = NULL)
}
\arguments{
\item{fname}{file name to be used to save out the data}

\item{vol}{MGH object (as from load.mgh)}
}
\description{
Save out an MGH file from memory
}
