% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QdecR_lm_24Nov2017.R
\name{calcFwhm}
\alias{calcFwhm}
\title{Compute smoothness of residuals for mc-z correction}
\usage{
calcFwhm(DIR, hemi, eres, mghBaseName, mask = NULL, target = "fsaverage",
  verbose = FALSE)
}
\arguments{
\item{DIR}{the output folder name, should have a trailing /}

\item{hemi}{lh or rh}

\item{eres}{the residual output from a vertex wise analysis}

\item{mghBaseName}{to help name the output dat file with the fwhm value}
}
\value{
fwhm numeric
}
\description{
Compute smoothness of residuals for mc-z correction
}
\examples{
fwhm <- calcFwhm('/export/local/scratch/test/', 'lh', '/export/local/scratch/test/test.eres.mgh', 'test.vertexWise')
}
