% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdecr_utility.R
\name{catprompt}
\alias{catprompt}
\title{\code{message} function with verbose option}
\usage{
catprompt(string, n = 60, base = "-", center = TRUE,
  addSpace = TRUE, start = 5, verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. If TRUE, \code{message} is called.}

\item{...}{Any input for \code{message}}
}
\description{
This is a wrapper for \code{message}. It calls \code{message} if \code{verbose}
is set to true.
}
