% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBM_functions.R
\name{fbm_functions_from_qdecr}
\alias{fbm_functions_from_qdecr}
\alias{fbm_row_mean}
\alias{fbm_col_mean}
\alias{fbm_row_is_0}
\alias{fbm_col_is_0}
\alias{fbm_row_sum}
\alias{fbm_col_sum}
\alias{fbm_row_sd}
\alias{fbm_col_sd}
\title{fbm functions from qdecr}
\usage{
fbm_row_mean(
  X,
  ncores = 1,
  row.ind = bigstatsr::rows_along(X),
  col.ind = bigstatsr::cols_along(X),
  row.mask = NULL,
  col.mask = NULL
)

fbm_col_mean(
  X,
  ncores = 1,
  row.ind = bigstatsr::rows_along(X),
  col.ind = bigstatsr::cols_along(X),
  row.mask = NULL,
  col.mask = NULL
)

fbm_row_is_0(
  X,
  ncores = 1,
  row.ind = bigstatsr::rows_along(X),
  col.ind = bigstatsr::cols_along(X),
  row.mask = NULL,
  col.mask = NULL
)

fbm_col_is_0(
  X,
  ncores = 1,
  row.ind = bigstatsr::rows_along(X),
  col.ind = bigstatsr::cols_along(X),
  row.mask = NULL,
  col.mask = NULL
)

fbm_row_sum(
  X,
  ncores = 1,
  row.ind = bigstatsr::rows_along(X),
  col.ind = bigstatsr::cols_along(X),
  row.mask = NULL,
  col.mask = NULL
)

fbm_col_sum(
  X,
  ncores = 1,
  row.ind = bigstatsr::rows_along(X),
  col.ind = bigstatsr::cols_along(X),
  row.mask = NULL,
  col.mask = NULL
)

fbm_row_sd(
  X,
  ncores = 1,
  row.ind = bigstatsr::rows_along(X),
  col.ind = bigstatsr::cols_along(X),
  row.mask = NULL,
  col.mask = NULL
)

fbm_col_sd(
  X,
  ncores = 1,
  row.ind = bigstatsr::rows_along(X),
  col.ind = bigstatsr::cols_along(X),
  row.mask = NULL,
  col.mask = NULL
)
}
\arguments{
\item{X}{the FBM (file-backed matrix) object}

\item{ncores}{the number of cores to be used}

\item{row.ind}{the rows to go over (by default, it goes over all rows)}

\item{col.ind}{the columns to go over (by default, it goes over all columns)}

\item{row.mask}{a separate mask for the rows (added because this was useful internally)}

\item{col.mask}{a separate mask for the columns (added because this was useful internally)}
}
\value{
an mgh object of the corresponding map
}
\description{
fbm functions from qdecr
}
