% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing.R
\name{formula.vw_fastlm}
\alias{formula.vw_fastlm}
\title{Grabs the formula from qdecr_fastlm output}
\usage{
\method{formula}{vw_fastlm}(x, ...)
}
\arguments{
\item{x}{The output object of a qdecr_fastlm call}

\item{...}{Further arguments for `formula`}
}
\description{
Grabs the formula from qdecr_fastlm output
}
\details{
See above.
}
