% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{freeview}
\alias{freeview}
\title{Opens Freeview}
\usage{
freeview(vw, stack = NULL, mask = TRUE)
}
\arguments{
\item{vw}{The output object of a qdecr call (e.g. qdecr_fastlm)}

\item{stack}{Either a numeric or a string for your variable (use `stacks(vw)`)}

\item{mask}{Logical; if TRUE, only the significant clusters are shown}
}
\description{
Opens Freeview to view the surface of a specific stack/contrast
}
\details{
This function directly calls Freeview so that you can view the surface of a specific
variable. As it opens Freeview directly, you will have access to all the functions
of Freeview by default. To continue using R, simply close Freeview.
}
