% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp2list.R
\name{imp2list}
\alias{imp2list}
\title{Convert imputation output to \code{list}}
\usage{
imp2list(obj)
}
\arguments{
\item{obj}{The imputation object that contains the imputed datasets}
}
\value{
Outputs a list of imputed datasets
}
\description{
Converts known imputation objects containing datasets to a \code{list} of
datasets
}
\details{
This function attempts to unify imputation object formats in supplying
datasets. This is done by extracting the imputed datasets from the imputation
object and assembling them into a \code{list}. If an unknown object is
supplied, the function will throw an error with the request to supply a list
of data frames instead.
}
\author{
Sander Lamballais, 2018.
}
