% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_mgh.R
\name{load.annot}
\alias{load.annot}
\title{Load an .annot file into memory}
\usage{
load.annot(input.file)
}
\arguments{
\item{input.file}{full path to the .annot file}
}
\value{
annot object with data and various header elements
}
\description{
Load an .annot file into memory
}
