% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QdecR_lm_24Nov2017.R
\name{mergeMgh}
\alias{mergeMgh}
\title{call the FreeSurfer tool mris_preproc to merge surface data into a single mgh file}
\usage{
mergeMgh(DIR, idVar, dataFrame, hemi, cacheIn, cacheOut, target = "fsaverage",
  verbose = FALSE)
}
\arguments{
\item{DIR}{the output folder name, should have a trailing /}

\item{idVar}{The id variable in your data frame that matches the id variable in the SUBJECTS_DIR}

\item{dataFrame}{The data frame with the list of subjects you want to merge}

\item{hemi}{lh or rh}

\item{cacheIn}{This is the root output name of the qcache recon-all. example: thickness.fwhm10.fsaverage}

\item{cacheOut}{This is the name of the file you want to save out in mgh format}

\item{target}{usually fsaverage, unless you made a custom template}

\item{verbose}{defaults to FALSE to supress loads out output from the merging process.}
}
\description{
call the FreeSurfer tool mris_preproc to merge surface data into a single mgh file
}
\examples{
mergeMgh('/export/local/scratch/tests', 'subjid', myData, 'lh', 'thickness.fwhm10.fsaverage', 'lh.thickness.fwhm10.fsaverage.mgh', verbose = verbose)

}
