% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing.R
\name{nobs.vw_fastlm}
\alias{nobs.vw_fastlm}
\title{Grabs the number of included participants from qdecr_fastlm output}
\usage{
\method{nobs}{vw_fastlm}(vw)
}
\arguments{
\item{vw}{The output object of a qdecr analysis call}
}
\description{
Grabs the number of included participants from qdecr_fastlm output
}
\details{
See above.
}
