% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing.R
\name{print.vw}
\alias{print.vw}
\title{Print for vw objects}
\usage{
\method{print}{vw}(x, ...)
}
\arguments{
\item{x}{The output object of a qdecr analysis call}

\item{...}{Further arguments for `print`}
}
\description{
Prints a nicely formatted output for vw objects
}
\details{
This is a standard print format for output of qdecr functions
}
