% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QdecR_lm_24Nov2017.R
\docType{package}
\name{QdecR-package}
\alias{QdecR}
\alias{QdecR-package}
\title{QdecR: Vertex-wise statistical analysis}
\description{
Vertex Wise analysis of MGH formatted Data
The main function to use is ...
}
\examples{
Put examples in here

}
\references{
Muetzel, et al, Bullying involvement and cortical morphology in children.
}
\author{
Ryan Muetzel & Sander Lamballais
}
