% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing.R
\name{qdecr_fwhm}
\alias{qdecr_fwhm}
\title{Grabs the estimated fwhm from qdecr_fastlm output}
\usage{
qdecr_fwhm(vw)
}
\arguments{
\item{vw}{The output object of a qdecr analysis call}
}
\description{
Grabs the estimated fwhm from qdecr_fastlm output
}
\details{
See above.
}
