% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdecr_load.R
\name{qdecr_load}
\alias{qdecr_load}
\title{Loads files associated with QDECR projects}
\usage{
qdecr_load(path, reload = FALSE)
}
\arguments{
\item{path}{Either the path to the directory, or a direct path to the associated .rds file}

\item{reload}{If TRUE, also reloads the vertex-wise matrix (which is HUGE!); default is FALSE}
}
\value{
Object of class "vw" that was generated by the respective analysis.
}
\description{
Loads files associated with QDECR projects
}
