% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdecr_read.R
\name{qdecr_read}
\alias{qdecr_read}
\alias{qdecr_read_ocn}
\alias{qdecr_read_ocn_mask}
\alias{qdecr_read_coef}
\alias{qdecr_read_p}
\alias{qdecr_read_t}
\alias{qdecr_read_se}
\title{qdecr_read function family}
\usage{
qdecr_read_ocn(vw, stack)

qdecr_read_ocn_mask(vw, stack)

qdecr_read_coef(vw, stack)

qdecr_read_p(vw, stack)

qdecr_read_t(vw, stack)

qdecr_read_se(vw, stack)
}
\arguments{
\item{vw}{Output object from a qdecr analysis}

\item{stack}{stack number of name of the stack of interest}
}
\value{
an mgh object of the corresponding map
}
\description{
qdecr_read function family
}
