% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdecr_save.R
\name{qdecr_save}
\alias{qdecr_save}
\title{Saves "vw" objects from the QDECR package}
\usage{
qdecr_save(vw, ...)
}
\arguments{
\item{vw}{Object of class "vw". Potentially other classes related to specific methods, e.g. "vw_fastlm"}
}
\description{
Saves "vw" objects from the QDECR package
}
