% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{qdecr_snap}
\alias{qdecr_snap}
\title{Opens Freeview to take snapshots}
\usage{
qdecr_snap(vw, stack = NULL, ext = ".tiff", zoom = 1,
  compose = TRUE, plot_brain = TRUE, save_plot = TRUE)
}
\arguments{
\item{vw}{The output object of a qdecr call (e.g. qdecr_fastlm)}

\item{stack}{Either a numeric or a string for your variable (use `stacks(vw)`)}

\item{ext}{Extension of the image files that will be stored on disk}

\item{zoom}{Float that determines how far the brains are zoomed in}

\item{compose}{Logical; if TRUE, a single compiled image will be made (requires Magick++)}

\item{plot_brain}{Logical; if TRUE, returns a graphical device with the composed images}

\item{save_plot}{Logical; if TRUE, saves the composed image}
}
\description{
Opens Freeview to take snapshots and compiles them into a plot
}
\details{
This function directly calls Freeview and generates files (by default .tiff)
and additionally composes the images and plots them. Magick++ is required
for the composing step (see www.qdecr.com).
}
