% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdecr_utility.R
\name{qdecr_update_path}
\alias{qdecr_update_path}
\title{Updates the internally stored paths in a vw object}
\usage{
qdecr_update_path(vw, dir_fshome, dir_subj,
  dir_project = dirname(vw$paths$rds), mask_path = NULL,
  qdecr_mask = TRUE, overwrite = TRUE)
}
\arguments{
\item{vw}{A `vw` object (QDECR output), e.g. from `qdecr_load`.}

\item{dir_fshome}{character; path to FREESURFER_HOME to be used.}

\item{dir_subj}{character; path to SUBJECTS_DIR to be used.}

\item{dir_project}{character; path to the project directory (defaults to the parent dir of the .rds file).}

\item{mask_path}{character; if specified and QDECR_mask == FALSE, this will become the mask path.}

\item{qdecr_mask}{logical (default = TRUE); should the QDECR mask be used as the mask?}

\item{overwrite}{logical (default = TRUE); should the project .rds file be overwritten?}
}
\value{
The vw object, but with updated paths.
}
\description{
Updates the internally stored paths in a vw object
}
