% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdecr_load.R
\name{reload}
\alias{reload}
\title{Reloads vertex-wise data for an "vw" object}
\usage{
reload(vw, verbose = TRUE)
}
\arguments{
\item{vw}{An object of class "vw"}

\item{verbose}{If TRUE, displays all information related to the reloading; default is TRUE}
}
\value{
Object of class "vw" where vw$mgh is populated.
}
\description{
Reloads vertex-wise data for an "vw" object
}
