% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QdecR_lm_24Nov2017.R
\name{run_lm_VertexWiseParRcpp_Y}
\alias{run_lm_VertexWiseParRcpp_Y}
\title{Run vertex wise analysis using mgh data
The "_Y" denotes that the vertex data are the outcome measure in the LM}
\usage{
run_lm_VertexWiseParRcpp_Y(idVar, dataFrame, hemi, outDir, x, covars = NULL,
  fitEq = NULL, mask = NULL, con.vec = NULL, target = "fsaverage",
  measure = "thickness", fwhm = ifelse(measure == "pial_lgi", 5, 10),
  mgh = NULL, mghBaseName, cleanUp = TRUE, cleanUpBM = TRUE,
  tmpDir = outDir, nCores = 2, clobber = FALSE, writeDesign = TRUE,
  verbose = FALSE)
}
\arguments{
\item{idVar}{The id variable in your data frame that matches the id variable in the SUBJECTS_DIR}

\item{dataFrame}{The data frame with the list of subjects you want to merge}

\item{hemi}{lh or rh}

\item{x}{variable of interest in data frame (e.g., age)}

\item{covars}{vector of variables you want added to the model}

\item{mask}{mgh file to mask analysis. default is to use the cortex label from the target}

\item{con.vec}{contrast vector if you want something other than just the positve contrast of x}

\item{target}{usually fsaverage, unless you made a custom template}

\item{measure}{thickness, pial_lgi, etc}

\item{fwhm}{smoothing kernel used on mgh data (defualt is 10)}

\item{nCores}{the number of cores you want to use (for Parallelization). Not sure how well this works with more than 10 cores)}

\item{verbose}{defaults to FALSE to supress loads out output from the merging process.}

\item{DIR}{the output folder name, should have a trailing /}
}
\description{
Run vertex wise analysis using mgh data
The "_Y" denotes that the vertex data are the outcome measure in the LM
}
