% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QdecR_lm_24Nov2017.R
\name{runMriSurfCluster}
\alias{runMriSurfCluster}
\title{Apply mc-z simulation correction to a log10p map}
\usage{
runMriSurfCluster(DIR, hemi, mghBaseName, fwhm, mask = NULL, cwpThr = 0.025,
  mczThr = NULL, csdSign = "abs", verbose = FALSE)
}
\arguments{
\item{DIR}{the output folder name, should have a trailing /}

\item{hemi}{lh or rh}

\item{mghBaseName}{to help name the output dat file with the fwhm value}

\item{fwhm}{(from calcFwhm)}

\item{mask}{mgh mask file, default is to just use the cortex label option}

\item{cwpThr}{cluster wise p thresh...default 0.025...i.e., 'two hemi'}

\item{mczThr}{mcz threshold...default is 3}

\item{csdSign}{sign for significance testing...abs (default), pos or neg}

\item{verbose}{allow terminal output}
}
\description{
Apply mc-z simulation correction to a log10p map
}
