% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_mgh.R
\name{save.mgh}
\alias{save.mgh}
\title{Save out an MGH file from memory}
\usage{
save.mgh(vol, fname)
}
\arguments{
\item{vol}{MGH object (as from load.mgh)}

\item{fname}{file name to be used to save out the data}
}
\description{
R translation of save_mgh.m (by Heath Pardoe, 09/12/2013)
}
