% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing.R
\name{stacks}
\alias{stacks}
\title{Shows vw stacks}
\usage{
stacks(vw)
}
\arguments{
\item{vw}{The output object of a qdecr analysis call}
}
\description{
Shows the names of the stacks of vw objects
}
\details{
This function shows the stack/contrast names for all the variables
within the qdecr analysis. This functions commonly used for looking
more specifically at variables.
}
