% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdecr_load.R
\name{unload}
\alias{unload}
\title{Removes vertex-wise data from an "vw" object}
\usage{
unload(vw)
}
\arguments{
\item{vw}{An object of class "vw" where vw$mgh is populated}
}
\value{
Object of class "vw" where vw$mgh is empty
}
\description{
Removes vertex-wise data from an "vw" object
}
