% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_mgh.R
\name{as_mgh}
\alias{as_mgh}
\title{Create an object that is structured like an mgh object}
\usage{
as_mgh(x, v = NULL, ndim1 = NULL, ndim2 = NULL, ndim3 = NULL,
  nframes, type, dof)
}
\arguments{
\item{x}{the vertex-wise values}

\item{v}{(to be added)}

\item{ndim1}{(to be added)}

\item{ndim2}{(to be added)}

\item{ndim3}{(to be added)}

\item{nframes}{(to be added)}

\item{type}{(to be added)}

\item{dof}{(to be added)}
}
\description{
Create an object that is structured like an mgh object
}
