% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_mgh.R
\name{bsfbm2mgh}
\alias{bsfbm2mgh}
\title{Save out an MGH file from a big matrix}
\usage{
bsfbm2mgh(fbm, fname, filter = NULL)
}
\arguments{
\item{fbm}{a file backed matrix}

\item{fname}{file name to be used to save out the data}

\item{filter}{a vector of indices of rows to include of fbm}
}
\description{
Modified version of save_mgh.m (by Heath Pardoe, 09/12/2013)
}
