% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing.R
\name{formula.vw_fastlm}
\alias{formula.vw_fastlm}
\title{Grabs the formula from qdecr_fastlm output}
\usage{
formula.vw_fastlm(vw)
}
\arguments{
\item{vw}{The output object of a qdecr_fastlm call}
}
\description{
Grabs the formula from qdecr_fastlm output
}
\details{
See above.
}
