% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{freeview}
\alias{freeview}
\title{Opens Freeview}
\usage{
freeview(vw, stack = NULL, type = c("coef", "se", "t", "p"), sig = TRUE, ...)
}
\arguments{
\item{vw}{The output object of a qdecr call (e.g. qdecr_fastlm)}

\item{stack}{Either a numeric or a string for your variable (use `stacks(vw)`)}

\item{type}{The type of map to plot (currently for OLS). Default is `coef`, others are `se`, `t` and `p`}

\item{sig}{Logical; if TRUE (default), only the significant clusters are shown}

\item{...}{any arguments that freeview (on command line) normally takes for manipulating surface files}
}
\description{
Opens Freeview to view the surface of a specific stack/contrast
}
\details{
This function directly calls Freeview so that you can view the surface of a specific
variable. As it opens Freeview directly, you will have access to all the functions
of Freeview by default. To continue using R, simply close Freeview.
}
