% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_mgh.R
\name{load.mgh}
\alias{load.mgh}
\title{Load an MGH file into memory}
\usage{
load.mgh(input.file)
}
\arguments{
\item{input.file}{full path to the mgh file}
}
\value{
mgh object with data and various header elements
}
\description{
Load an MGH file into memory
}
