% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdecr.R
\name{qdecr}
\alias{qdecr}
\title{Run vertex-wise analyses}
\usage{
qdecr(id, data, vertex = "qdecr", margs = NULL,
  model = c("stats::lm", "QDECR::megha", "RcppEigen::fastLm",
  "stats::glm", "survival::coxph", "default"), target = "fsaverage",
  hemi = c("lh", "rh"), measure = c("thickness", "area", "area.pial",
  "curv", "jacobian_white", "pial", "pial_lgi", "sulc", "volume",
  "w-g.pct.mgh", "white.H", "white.K"), fwhm = ifelse(measure ==
  "pial_lgi", 5, 10), mcz_thr = 30, mgh = NULL, mask = NULL, project,
  dir_subj = Sys.getenv("SUBJECTS_DIR"),
  dir_fshome = Sys.getenv("FREESURFER_HOME"), dir_tmp = dir_out,
  dir_out, dir_out_tree = TRUE, clean_up = TRUE, clean_up_bm = TRUE,
  clobber = FALSE, verbose = TRUE, debug = FALSE, n_cores = 1,
  sander = FALSE)
}
\arguments{
\item{a}{test}

\item{b}{test}

\item{c}{test}

\item{d}{test}

\item{e}{test}

\item{f}{test}

\item{g}{test}

\item{h}{test}
}
\value{
out
}
\description{
Runs vertex-wise analyses using a variety of statistical models
}
\details{
This function is the worker function of the \code{QDECR} package. It allows
.mgh format data as input and allows statistical analyses per vertex. A variety
of statistical models have been implemented, such as linear regression and 
GCTA.
}
