% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdecr_fastlm.R
\name{qdecr_fastlm}
\alias{qdecr_fastlm}
\title{Vertex-wise linear regression (based on RcppEigen)}
\usage{
qdecr_fastlm(formula, data, id, hemi, dir_out = getwd(), project,
  n_cores = 1, target = "fsaverage", mcz_thr = 30, mgh = NULL,
  mask = NULL, dir_subj = Sys.getenv("SUBJECTS_DIR"),
  dir_fshome = Sys.getenv("FREESURFER_HOME"), dir_tmp = dir_out,
  dir_out_tree = TRUE, clean_up_bm = TRUE, clean_up = TRUE,
  clobber = FALSE, verbose = TRUE, save = TRUE, save_data = TRUE,
  debug = FALSE)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. See `?lm`.}

\item{data}{a required argument that contains a data frame, a list of data frames or an imputed object that is supported by the `imp2list` function (mice, mi, etc.).}

\item{id}{the name of the id variable that matches the dataset to the Freesurfer output}

\item{hemi}{"lh" or "rh"}

\item{dir_out}{the directory where to save the data to (defaults to the current directory)}

\item{project}{the base name you want to assign to the output files}

\item{n_cores}{the number of cores to be used}

\item{target}{the target template (usually "fsaverage")}

\item{mcz_thr}{the Monte Carlo simulation threshold times 10 (13 = 0.05, 20 = 0.01, 23 = 0.005 30 = 0.001, etc..)}

\item{mgh}{NOT IMPLEMENTED; path to existing merged mgh file, default is NULL}

\item{mask}{mgh file to mask analysis; default is to use the cortex label from the target}

\item{dir_subj}{directory contain the surface-based maps (mgh files); defaults to SUBJECTS_DIR}

\item{dir_fshome}{Freesurfer directory; defaults to FREESURFER_HOME}

\item{dir_tmp}{directory to store the temporary big matrices; useful for shared memory; defaults to `dir_out`}

\item{dir_out_tree}{if TRUE, creates a dir_out/project directory. If FALSE, all output is placed directory into dir_out}

\item{clean_up_bm}{if TRUE, cleans all big matrices (.bk) that were generated in dir_tmp}

\item{clean_up}{NOT IMPLEMENTED; will be used for setting cleaning of other files}

\item{clobber}{if TRUE, ignores already existing directories and writes over them; if FALSE, stops and warns user that a given directory already exists}

\item{verbose}{if TRUE, writes out standard log; if FALSE, no output is generated}

\item{save}{if TRUE, saves the output to a .fst file}

\item{save_data}{if TRUE, includes the raw data + design matrices in the .fst file}

\item{debug}{NOT IMPLEMENTED; will output the maximal log to allow for easy debugging}
}
\value{
returns an object of classes "vw_fastlm" and "vw".
}
\description{
Vertex-wise linear regression (based on RcppEigen)
}
