% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_mgh.R
\name{qdecr_prep_mgh}
\alias{qdecr_prep_mgh}
\title{Combine all MGH files into one object}
\usage{
qdecr_prep_mgh(input_path, files_list = list.files(input_path), mask,
  hemi, measure, fwhmc, target, n_cores, dir_tmp, project, backing,
  verbose)
}
\arguments{
\item{input.list}{path to parent directory}

\item{files.list}{file names within `input.list`}
}
\value{
matrix with vertex data (columns) for each file (rows)
}
\description{
Combine all MGH files into one object
}
